# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "iRestora POS API",
        "description": "API Documentation untuk Sistem POS F&B iRestora",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/orders/{orderId}/payments": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Tambah pembayaran ke order",
                "operationId": "122a7c739ae1d2dc13a90c5a7bfc3482",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "method",
                                    "amount"
                                ],
                                "properties": {
                                    "method": {
                                        "type": "string",
                                        "enum": [
                                            "cash",
                                            "qris",
                                            "card",
                                            "other"
                                        ]
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "reference_number": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pembayaran ditambahkan, order status diupdate ke paid jika lunas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{orderId}/void": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Void/batal order (butuh approval PIN supervisor)",
                "operationId": "8d977415f7e1a9554ed0285e90410215",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reason",
                                    "approved_by_pin"
                                ],
                                "properties": {
                                    "reason": {
                                        "type": "string"
                                    },
                                    "approved_by_pin": {
                                        "description": "PIN supervisor/manager",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order dibatalkan, tercatat di audit_logs"
                    },
                    "403": {
                        "description": "PIN approval tidak valid / tidak memiliki izin"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/shifts/active": {
            "get": {
                "tags": [
                    "Shifts"
                ],
                "summary": "Get active shift for an outlet",
                "operationId": "76b12824af35a5b6c7f08ff4899c0afa",
                "parameters": [
                    {
                        "name": "outlet_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Active shift or null"
                    },
                    "404": {
                        "description": "No active shift"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/shifts/open": {
            "post": {
                "tags": [
                    "Shifts"
                ],
                "summary": "Buka shift kasir",
                "operationId": "4040ed223e9694aaa548471c4d86f8b6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "outlet_id",
                                    "opening_cash"
                                ],
                                "properties": {
                                    "outlet_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "opening_cash": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Shift dibuka"
                    },
                    "409": {
                        "description": "Sudah ada shift aktif untuk outlet ini"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/shifts/{shiftId}/close": {
            "post": {
                "tags": [
                    "Shifts"
                ],
                "summary": "Tutup shift kasir, hitung selisih kas",
                "operationId": "c7f75440df9ccdcafb37abdf0645fffd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "closing_cash_actual"
                                ],
                                "properties": {
                                    "closing_cash_actual": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Shift ditutup"
                    },
                    "404": {
                        "description": "Shift tidak ditemukan"
                    },
                    "422": {
                        "description": "Shift sudah ditutup atau bukan milik user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}